Poprawy

+++ database/migrations/2021_10_22_081510_create_employees_table.php
@@ -21,7 +21,7 @@ class CreateEmployeesTable extends Migration
             $table->bigInteger('phone');
             $table->enum('gender',['M','F']);
             $table->date('hire_date');
-            $table->bigInteger('salary');
+            $table->double('salary');
         });
     }
